WIP: Refactor create transaction
