+++ src/api/http.go
@@ -333,26 +333,50 @@ func newServerMux(c muxConfig, gateway Gatewayer) *http.ServeMux {
 		return handler
 	}
 
-	forAPISet := func(f http.HandlerFunc, apiNames []string) http.HandlerFunc {
-		if len(apiNames) == 0 {
-			logger.Panic("apiNames should not be empty")
+	forMethodAPISets := func(apiVersion string, f http.HandlerFunc, methodsAPISets map[string][]string) http.HandlerFunc {
+		if len(methodsAPISets) == 0 {
+			logger.Panic("methodsAPISets should not be empty")
 		}
 
-		isEnabled := false
+		switch apiVersion {
+		case apiVersion1, apiVersion2:
+		default:
+			logger.Panicf("Invalid API version %q", apiVersion)
+		}
 
-		for _, k := range apiNames {
-			if _, ok := c.enabledAPISets[k]; ok {
-				isEnabled = true
-				break
-			}
+		isEnabled := func(method string) bool {
 		}
 
 		return func(w http.ResponseWriter, r *http.Request) {
-			if isEnabled {
-				f(w, r)
-			} else {
-				wh.Error403(w, "Endpoint is disabled")
+			apiSets := methodsAPISets[method]
+
+			// If no API sets are specified for a given method, return 405 Method Not Allowed
+			if len(apiSets) == 0 {
+				switch apiVersion {
+				case apiVersion1:
+					wh.Error405(w, "")
+				case apiVersion2:
+					resp := NewHTTPErrorResponse(http.StatusMethodNotAllowed, "")
+					writeHTTPResponse(w, resp)
+				}
+				return
 			}
+
+			for _, k := range apiSets {
+				if _, ok := c.enabledAPISets[k]; ok {
+					f(w, r)
+					return
+				}
+			}
+
+			switch apiVersion {
+			case apiVersion1:
+				wh.Error403(w, "Endpoint is disabled")
+			case apiVersion2:
+				resp := NewHTTPErrorResponse(http.StatusForbidden, "Endpoint is disabled")
+				writeHTTPResponse(w, resp)
+			}
+			return
 		}
 	}
 
@@ -374,16 +398,20 @@ func newServerMux(c muxConfig, gateway Gatewayer) *http.ServeMux {
 		mux.Handle(endpoint, handler)
 	}
 
-	webHandler := func(apiVersion, endpoint string, handler http.Handler) {
+	webHandler := func(apiVersion, endpoint string, handler http.Handler, methodAPISets map[string][]string) {
+		if methodAPISets == nil /* explicitly check nil, caller should not pass empty initialized map */ {
+			handler = forMethodAPISets(apiVersion, handler, methodAPISets)
+		}
+
 		webHandlerWithOptionals(apiVersion, endpoint, handler, true, !c.disableHeaderCheck)
 	}
 
-	webHandlerV1 := func(endpoint string, handler http.Handler) {
-		webHandler(apiVersion1, "/api/v1"+endpoint, handler)
+	webHandlerV1 := func(endpoint string, handler http.Handler, methodAPISets map[string][]string) {
+		webHandler(apiVersion1, "/api/v1"+endpoint, handler, methodAPISets)
 	}
 
-	webHandlerV2 := func(endpoint string, handler http.Handler) {
-		webHandler(apiVersion2, "/api/v2"+endpoint, handler)
+	webHandlerV2 := func(endpoint string, handler http.Handler, methodAPISets map[string][]string) {
+		webHandler(apiVersion2, "/api/v2"+endpoint, handler, methodAPISets)
 	}
 
 	indexHandler := newIndexHandler(c.appLoc, c.enableGUI)
@@ -420,71 +448,165 @@ func newServerMux(c muxConfig, gateway Gatewayer) *http.ServeMux {
 	csrfHandlerV1("/csrf", getCSRFToken(c.disableCSRF)) // csrf is always available, regardless of the API set
 
 	// Status endpoints
-	webHandlerV1("/version", versionHandler(c.health.BuildInfo)) // version is always available, regardless of the API set
-	webHandlerV1("/health", forAPISet(healthHandler(c, gateway), []string{EndpointsRead, EndpointsStatus}))
+	webHandlerV1("/version", versionHandler(c.health.BuildInfo), nil) // version is always available, regardless of the API set
+	webHandlerV1("/health", healthHandler(c, gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
 
 	// Wallet endpoints
-	webHandlerV1("/wallet", forAPISet(walletHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/create", forAPISet(walletCreateHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/newAddress", forAPISet(walletNewAddressesHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/balance", forAPISet(walletBalanceHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/transaction", forAPISet(createTransactionHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV2("/wallet/transaction/sign", forAPISet(walletSignTransactionHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/transactions", forAPISet(walletTransactionsHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/update", forAPISet(walletUpdateHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallets", forAPISet(walletsHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallets/folderName", forAPISet(walletFolderHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/newSeed", forAPISet(newSeedHandler(), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/seed", forAPISet(walletSeedHandler(gateway), []string{EndpointsInsecureWalletSeed}))
-	webHandlerV2("/wallet/seed/verify", forAPISet(walletVerifySeedHandler, []string{EndpointsWallet}))
+	webHandlerV1("/wallet", walletHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/create", walletCreateHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/newAddress", walletNewAddressesHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/balance", walletBalanceHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/transaction", walletCreateTransactionHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV2("/wallet/transaction/sign", walletSignTransactionHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/transactions", walletTransactionsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/update", walletUpdateHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallets", walletsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallets/folderName", walletFolderHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/newSeed", newSeedHandler(), map[string][]string{
+		http.MethodGet: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/seed", walletSeedHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsInsecureWalletSeed},
+	})
+	webHandlerV2("/wallet/seed/verify", walletVerifySeedHandler, map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
 
-	webHandlerV1("/wallet/unload", forAPISet(walletUnloadHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/encrypt", forAPISet(walletEncryptHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV1("/wallet/decrypt", forAPISet(walletDecryptHandler(gateway), []string{EndpointsWallet}))
-	webHandlerV2("/wallet/recover", forAPISet(walletRecoverHandler(gateway), []string{EndpointsWallet}))
+	webHandlerV1("/wallet/unload", walletUnloadHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/encrypt", walletEncryptHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV1("/wallet/decrypt", walletDecryptHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
+	webHandlerV2("/wallet/recover", walletRecoverHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsWallet},
+	})
 
 	// Blockchain interface
-	webHandlerV1("/blockchain/metadata", forAPISet(blockchainMetadataHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
-	webHandlerV1("/blockchain/progress", forAPISet(blockchainProgressHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
-	webHandlerV1("/block", forAPISet(blockHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/blocks", forAPISet(blocksHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/last_blocks", forAPISet(lastBlocksHandler(gateway), []string{EndpointsRead}))
+	webHandlerV1("/blockchain/metadata", blockchainMetadataHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
+	webHandlerV1("/blockchain/progress", blockchainProgressHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
+	webHandlerV1("/block", blockHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/blocks", blocksHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/last_blocks", lastBlocksHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
 
 	// Network stats endpoints
-	webHandlerV1("/network/connection", forAPISet(connectionHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
-	webHandlerV1("/network/connections", forAPISet(connectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
-	webHandlerV1("/network/defaultConnections", forAPISet(defaultConnectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
-	webHandlerV1("/network/connections/trust", forAPISet(trustConnectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
-	webHandlerV1("/network/connections/exchange", forAPISet(exchgConnectionsHandler(gateway), []string{EndpointsRead, EndpointsStatus}))
+	webHandlerV1("/network/connection", connectionHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
+	webHandlerV1("/network/connections", connectionsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
+	webHandlerV1("/network/defaultConnections", defaultConnectionsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
+	webHandlerV1("/network/connections/trust", trustConnectionsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
+	webHandlerV1("/network/connections/exchange", exchgConnectionsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead, EndpointsStatus},
+	})
 
 	// Network admin endpoints
-	webHandlerV1("/network/connection/disconnect", forAPISet(disconnectHandler(gateway), []string{EndpointsNetCtrl}))
+	webHandlerV1("/network/connection/disconnect", disconnectHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsNetCtrl},
+	})
 
 	// Transaction related endpoints
-	webHandlerV1("/pendingTxs", forAPISet(pendingTxnsHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/transaction", forAPISet(transactionHandler(gateway), []string{EndpointsRead}))
-	webHandlerV2("/transaction/verify", forAPISet(verifyTxnHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/transactions", forAPISet(transactionsHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/injectTransaction", forAPISet(injectTransactionHandler(gateway), []string{EndpointsTransaction, EndpointsWallet}))
-	webHandlerV1("/resendUnconfirmedTxns", forAPISet(resendUnconfirmedTxnsHandler(gateway), []string{EndpointsTransaction}))
-	webHandlerV1("/rawtx", forAPISet(rawTxnHandler(gateway), []string{EndpointsRead}))
+	webHandlerV1("/pendingTxs", pendingTxnsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/transaction", transactionHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV2("/transaction", transactionHandlerV2(gateway), map[string][]string{
+		// http.MethodGet:  []string{EndpointsRead},
+		http.MethodPost: []string{EndpointsTransaction},
+	})
+	webHandlerV2("/transaction/verify", verifyTxnHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsRead},
+	})
+	webHandlerV1("/transactions", transactionsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/injectTransaction", injectTransactionHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsTransaction, EndpointsWallet},
+	})
+	webHandlerV1("/resendUnconfirmedTxns", resendUnconfirmedTxnsHandler(gateway), map[string][]string{
+		http.MethodPost: []string{EndpointsTransaction, EndpointsWallet},
+	})
+	webHandlerV1("/rawtx", rawTxnHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
 
 	// Unspent output related endpoints
-	webHandlerV1("/outputs", forAPISet(outputsHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/balance", forAPISet(balanceHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/uxout", forAPISet(uxOutHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/address_uxouts", forAPISet(addrUxOutsHandler(gateway), []string{EndpointsRead}))
+	webHandlerV1("/outputs", outputsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/balance", balanceHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/uxout", uxOutHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/address_uxouts", addrUxOutsHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
 
 	// golang process internal metrics for Prometheus
-	webHandlerV2("/metrics", forAPISet(promhttp.Handler().(http.HandlerFunc), []string{EndpointsPrometheus}))
+	webHandlerV2("/metrics", promhttp.Handler().(http.HandlerFunc), map[string][]string{
+		http.MethodGet: []string{EndpointsPrometheus},
+	})
 
 	// Address related endpoints
-	webHandlerV2("/address/verify", forAPISet(addressVerifyHandler, []string{EndpointsRead}))
+	webHandlerV2("/address/verify", addressVerifyHandler, map[string][]string{
+		http.MethodPost: []string{EndpointsRead},
+	})
 
 	// Explorer endpoints
-	webHandlerV1("/coinSupply", forAPISet(coinSupplyHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/richlist", forAPISet(richlistHandler(gateway), []string{EndpointsRead}))
-	webHandlerV1("/addresscount", forAPISet(addressCountHandler(gateway), []string{EndpointsRead}))
+	webHandlerV1("/coinSupply", coinSupplyHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/richlist", richlistHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
+	webHandlerV1("/addresscount", addressCountHandler(gateway), map[string][]string{
+		http.MethodGet: []string{EndpointsRead},
+	})
 
 	return mux
 }
