+++ src/visor/visor.go
@@ -2275,122 +2275,6 @@ func (vs *Visor) AddressCount() (uint64, error) {
 	return count, nil
 }
 
-func (vs *Visor) getCreateTransactionAuxs(tx *dbutil.Tx, params wallet.CreateTransactionParams, allAddrs []cipher.Address) (coin.AddressUxOuts, error) {
-	allAddrsMap := make(map[cipher.Address]struct{}, len(allAddrs))
-	for _, a := range allAddrs {
-		allAddrsMap[a] = struct{}{}
-	}
-
-	var auxs coin.AddressUxOuts
-	if len(params.Wallet.UxOuts) != 0 {
-		// Check if any of the outputs are in an unconfirmed spend
-		hashesMap := make(map[cipher.SHA256]struct{}, len(params.Wallet.UxOuts))
-		for _, h := range params.Wallet.UxOuts {
-			hashesMap[h] = struct{}{}
-		}
-
-		// Get all unconfirmed spending uxouts
-		unconfirmedTxns, err := vs.Unconfirmed.AllRawTransactions(tx)
-		if err != nil {
-			return nil, err
-		}
-
-		var unconfirmedSpends []cipher.SHA256
-		for _, txn := range unconfirmedTxns {
-			unconfirmedSpends = append(unconfirmedSpends, txn.In...)
-		}
-
-		if params.IgnoreUnconfirmed {
-			// Filter unconfirmed spends
-			prevLen := len(hashesMap)
-			for _, h := range unconfirmedSpends {
-				delete(hashesMap, h)
-			}
-
-			if prevLen != len(hashesMap) {
-				params.Wallet.UxOuts = make([]cipher.SHA256, 0, len(hashesMap))
-				for h := range hashesMap {
-					params.Wallet.UxOuts = append(params.Wallet.UxOuts, h)
-				}
-			}
-		} else {
-			for _, h := range unconfirmedSpends {
-				if _, ok := hashesMap[h]; ok {
-					return nil, wallet.ErrSpendingUnconfirmed
-				}
-			}
-		}
-
-		// Retrieve the uxouts from the pool.
-		// An error is returned if any do not exist
-		uxouts, err := vs.Blockchain.Unspent().GetArray(tx, params.Wallet.UxOuts)
-		if err != nil {
-			return nil, err
-		}
-
-		// Build coin.AddressUxOuts map, and check that the address is in the wallets
-		auxs = make(coin.AddressUxOuts)
-		for _, o := range uxouts {
-			if _, ok := allAddrsMap[o.Body.Address]; !ok {
-				return nil, wallet.ErrUnknownUxOut
-			}
-			auxs[o.Body.Address] = append(auxs[o.Body.Address], o)
-		}
-
-	} else {
-		addrs := params.Wallet.Addresses
-		if len(addrs) == 0 {
-			addrs = allAddrs
-		} else {
-			// Check that requested addresses are in the wallet
-			for _, a := range addrs {
-				if _, ok := allAddrsMap[a]; !ok {
-					return nil, wallet.ErrUnknownAddress
-				}
-			}
-		}
-
-		// Get unspent outputs, while checking that there are no unconfirmed outputs
-		var err error
-		auxs, err = vs.getUnspentsForSpending(tx, addrs, params.IgnoreUnconfirmed)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return auxs, nil
-}
-
-// getUnspentsForSpending returns the unspent outputs for a set of addresses,
-// but returns an error if any of the unspents are in the unconfirmed outputs pool
-func (vs *Visor) getUnspentsForSpending(tx *dbutil.Tx, addrs []cipher.Address, ignoredUnconfirmed bool) (coin.AddressUxOuts, error) {
-	unconfirmedAuxs, err := vs.unconfirmedSpendsOfAddresses(tx, addrs)
-	if err != nil {
-		err = fmt.Errorf("UnconfirmedSpendsOfAddresses failed: %v", err)
-		return nil, err
-	}
-
-	if !ignoredUnconfirmed {
-		// Check that this is not trying to spend unconfirmed outputs
-		if len(unconfirmedAuxs) > 0 {
-			return nil, wallet.ErrSpendingUnconfirmed
-		}
-	}
-
-	auxs, err := vs.Blockchain.Unspent().GetUnspentsOfAddrs(tx, addrs)
-	if err != nil {
-		err = fmt.Errorf("GetUnspentsOfAddrs failed: %v", err)
-		return nil, err
-	}
-
-	// Filter unconfirmed
-	if ignoredUnconfirmed && len(unconfirmedAuxs) > 0 {
-		auxs = auxs.Sub(unconfirmedAuxs)
-	}
-
-	return auxs, nil
-}
-
 // GetVerboseTransactionsForAddress returns verbose transaction data for a given address
 func (vs *Visor) GetVerboseTransactionsForAddress(a cipher.Address) ([]Transaction, [][]TransactionInput, error) {
 	var txns []Transaction
