+++ src/wallet/wallet.go
@@ -4,14 +4,11 @@ Package wallet implements wallets and the wallet database service
 package wallet
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"io/ioutil"
-	"math/big"
 	"os"
 	"path/filepath"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
@@ -19,14 +16,8 @@ import (
 	"encoding/hex"
 
 	"github.com/skycoin/skycoin/src/cipher"
-	"github.com/skycoin/skycoin/src/coin"
-	"github.com/skycoin/skycoin/src/params"
 
-	"github.com/shopspring/decimal"
-
-	"github.com/skycoin/skycoin/src/util/fee"
 	"github.com/skycoin/skycoin/src/util/logging"
-	"github.com/skycoin/skycoin/src/util/mathutil"
 )
 
 // Error wraps wallet-related errors.
@@ -95,12 +86,6 @@ var (
 	ErrUnknownUxOut = NewError(errors.New("uxout is not owned by any address in the wallet"))
 	// ErrNoUnspents is returned if a wallet has no unspents to spend
 	ErrNoUnspents = NewError(errors.New("no unspents to spend"))
-	// ErrNullChangeAddress ChangeAddress must not be the null address
-	ErrNullChangeAddress = NewError(errors.New("ChangeAddress must not be the null address"))
-	// ErrMissingTo To is required
-	ErrMissingTo = NewError(errors.New("To is required"))
-	// ErrZeroCoinsTo To.Coins must not be zero
-	ErrZeroCoinsTo = NewError(errors.New("To.Coins must not be zero"))
 	// ErrWalletRecoverSeedWrong is returned if the seed does not match the specified wallet when recovering
 	ErrWalletRecoverSeedWrong = NewError(errors.New("wallet recovery seed is wrong"))
 	// ErrNilBalanceGetter is returned if Options.ScanN > 0 but a nil BalanceGetter was provided
@@ -109,38 +94,6 @@ var (
 	ErrWalletNotDeterministic = NewError(errors.New("wallet type is not deterministic"))
 	// ErrInvalidCoinType is returned for invalid coin types
 	ErrInvalidCoinType = NewError(errors.New("invalid coin type"))
-	// ErrNullAddressTo To.Address must not be the null address
-	ErrNullAddressTo = NewError(errors.New("To.Address must not be the null address"))
-	// ErrDuplicateTo To contains duplicate values
-	ErrDuplicateTo = NewError(errors.New("To contains duplicate values"))
-	// ErrMissingWalletID Wallet.ID is required
-	ErrMissingWalletID = NewError(errors.New("Wallet.ID is required"))
-	// ErrIncludesNullAddress Wallet.Addresses must not contain the null address
-	ErrIncludesNullAddress = NewError(errors.New("Wallet.Addresses must not contain the null address"))
-	// ErrDuplicateAddresses Wallet.Addresses contains duplicate values
-	ErrDuplicateAddresses = NewError(errors.New("Wallet.Addresses contains duplicate values"))
-	// ErrZeroToHoursAuto To.Hours must be zero for auto type hours selection
-	ErrZeroToHoursAuto = NewError(errors.New("To.Hours must be zero for auto type hours selection"))
-	// ErrMissingModeAuto HoursSelection.Mode is required for auto type hours selection
-	ErrMissingModeAuto = NewError(errors.New("HoursSelection.Mode is required for auto type hours selection"))
-	// ErrInvalidHoursSelMode Invalid HoursSelection.Mode
-	ErrInvalidHoursSelMode = NewError(errors.New("Invalid HoursSelection.Mode"))
-	// ErrInvalidModeManual HoursSelection.Mode cannot be used for manual type hours selection
-	ErrInvalidModeManual = NewError(errors.New("HoursSelection.Mode cannot be used for manual type hours selection"))
-	// ErrInvalidHoursSelType Invalid HoursSelection.Type
-	ErrInvalidHoursSelType = NewError(errors.New("Invalid HoursSelection.Type"))
-	// ErrMissingShareFactor HoursSelection.ShareFactor must be set for share mode
-	ErrMissingShareFactor = NewError(errors.New("HoursSelection.ShareFactor must be set for share mode"))
-	// ErrInvalidShareFactor HoursSelection.ShareFactor can only be used for share mode
-	ErrInvalidShareFactor = NewError(errors.New("HoursSelection.ShareFactor can only be used for share mode"))
-	// ErrShareFactorOutOfRange HoursSelection.ShareFactor must be >= 0 and <= 1
-	ErrShareFactorOutOfRange = NewError(errors.New("HoursSelection.ShareFactor must be >= 0 and <= 1"))
-	// ErrWalletParamsConflict Wallet.UxOuts and Wallet.Addresses cannot be combined
-	ErrWalletParamsConflict = NewError(errors.New("Wallet.UxOuts and Wallet.Addresses cannot be combined"))
-	// ErrDuplicateUxOuts Wallet.UxOuts contains duplicate values
-	ErrDuplicateUxOuts = NewError(errors.New("Wallet.UxOuts contains duplicate values"))
-	// ErrUnknownWalletID params.Wallet.ID does not match wallet
-	ErrUnknownWalletID = NewError(errors.New("params.Wallet.ID does not match wallet"))
 )
 
 const (
@@ -188,150 +141,6 @@ const (
 
 // CoinType represents the wallet coin type
 type CoinType string
-
-const (
-	// HoursSelectionTypeManual is used to specify manual hours selection in advanced spend
-	HoursSelectionTypeManual = "manual"
-	// HoursSelectionTypeAuto is used to specify automatic hours selection in advanced spend
-	HoursSelectionTypeAuto = "auto"
-
-	// HoursSelectionModeShare will distribute coin hours equally amongst destinations
-	HoursSelectionModeShare = "share"
-)
-
-// HoursSelection defines options for hours distribution
-type HoursSelection struct {
-	Type        string
-	Mode        string
-	ShareFactor *decimal.Decimal
-}
-
-// CreateTransactionWalletParams defines a wallet to spend from and optionally which addresses in the wallet
-type CreateTransactionWalletParams struct {
-	ID        string
-	UxOuts    []cipher.SHA256
-	Addresses []cipher.Address
-	Password  []byte
-}
-
-// CreateTransactionParams defines control parameters for transaction construction
-type CreateTransactionParams struct {
-	Unsigned          bool
-	IgnoreUnconfirmed bool
-	HoursSelection    HoursSelection
-	Wallet            CreateTransactionWalletParams
-	ChangeAddress     *cipher.Address
-	To                []coin.TransactionOutput
-}
-
-// Validate validates CreateTransactionParams
-func (c CreateTransactionParams) Validate() error {
-	if c.ChangeAddress != nil && c.ChangeAddress.Null() {
-		return ErrNullChangeAddress
-	}
-
-	if len(c.To) == 0 {
-		return ErrMissingTo
-	}
-
-	for _, to := range c.To {
-		if to.Coins == 0 {
-			return ErrZeroCoinsTo
-		}
-
-		if to.Address.Null() {
-			return ErrNullAddressTo
-		}
-	}
-
-	// Check for duplicate outputs, a transaction can't have outputs with
-	// the same (address, coins, hours)
-	// Auto mode would distribute hours to the outputs and could hypothetically
-	// avoid assigning duplicate hours in many cases, but the complexity for doing
-	// so is very high, so also reject duplicate (address, coins) for auto mode.
-	outputs := make(map[coin.TransactionOutput]struct{}, len(c.To))
-	for _, to := range c.To {
-		outputs[to] = struct{}{}
-	}
-
-	if len(outputs) != len(c.To) {
-		return ErrDuplicateTo
-	}
-
-	if c.Wallet.ID == "" {
-		return ErrMissingWalletID
-	}
-
-	addressMap := make(map[cipher.Address]struct{}, len(c.Wallet.Addresses))
-	for _, a := range c.Wallet.Addresses {
-		if a.Null() {
-			return ErrIncludesNullAddress
-		}
-
-		addressMap[a] = struct{}{}
-	}
-
-	if len(addressMap) != len(c.Wallet.Addresses) {
-		return ErrDuplicateAddresses
-	}
-
-	switch c.HoursSelection.Type {
-	case HoursSelectionTypeAuto:
-		for _, to := range c.To {
-			if to.Hours != 0 {
-				return ErrZeroToHoursAuto
-			}
-		}
-
-		switch c.HoursSelection.Mode {
-		case HoursSelectionModeShare:
-		case "":
-			return ErrMissingModeAuto
-		default:
-			return ErrInvalidHoursSelMode
-		}
-
-	case HoursSelectionTypeManual:
-		if c.HoursSelection.Mode != "" {
-			return ErrInvalidModeManual
-		}
-
-	default:
-		return ErrInvalidHoursSelType
-	}
-
-	if c.HoursSelection.ShareFactor == nil {
-		if c.HoursSelection.Mode == HoursSelectionModeShare {
-			return ErrMissingShareFactor
-		}
-	} else {
-		if c.HoursSelection.Mode != HoursSelectionModeShare {
-			return ErrInvalidShareFactor
-		}
-
-		zero := decimal.New(0, 0)
-		one := decimal.New(1, 0)
-		if c.HoursSelection.ShareFactor.LessThan(zero) || c.HoursSelection.ShareFactor.GreaterThan(one) {
-			return ErrShareFactorOutOfRange
-		}
-	}
-
-	if len(c.Wallet.UxOuts) != 0 && len(c.Wallet.Addresses) != 0 {
-		return ErrWalletParamsConflict
-	}
-
-	// Check for duplicate spending uxouts
-	uxouts := make(map[cipher.SHA256]struct{}, len(c.Wallet.UxOuts))
-	for _, o := range c.Wallet.UxOuts {
-		uxouts[o] = struct{}{}
-	}
-
-	if len(uxouts) != len(c.Wallet.UxOuts) {
-		return ErrDuplicateUxOuts
-	}
-
-	return nil
-}
 
 // NewWalletFilename generates a filename from the current time and random bytes
 func NewWalletFilename() string {
@@ -1112,6 +921,18 @@ func (w *Wallet) GetEntry(a cipher.Address) (Entry, bool) {
 	return Entry{}, false
 }
 
+// HasEntry returns true if the wallet has an Entry with a given cipher.Address.
+func (w *Wallet) HasEntry(a cipher.Address) bool {
+	// This doesn't use GetEntry() to avoid copying an Entry in the return value,
+	// which may contain a secret key
+	for _, e := range w.Entries {
+		if e.SkycoinAddress() == a {
+			return true
+		}
+	}
+	return false
+}
+
 // AddEntry adds new entry
 func (w *Wallet) AddEntry(entry Entry) error {
 	// dup check
@@ -1137,977 +958,4 @@ func (w *Wallet) clone() *Wallet {
 	wlt.Entries = append(wlt.Entries, w.Entries...)
 
 	return &wlt
-}
-
-func validateSignIndexes(x []int, uxOuts []coin.UxOut) error {
-	if len(x) > len(uxOuts) {
-		return errors.New("Number of signature indexes exceeds number of inputs")
-	}
-
-	for _, i := range x {
-		if i >= len(uxOuts) || i < 0 {
-			return errors.New("Signature index out of range")
-		}
-	}
-
-	m := make(map[int]struct{}, len(x))
-	for _, i := range x {
-		if _, ok := m[i]; ok {
-			return errors.New("Duplicate value in signature indexes")
-		}
-		m[i] = struct{}{}
-	}
-
-	return nil
-}
-
-func copyTransaction(txn *coin.Transaction) *coin.Transaction {
-	txnHash := txn.Hash()
-	txnInnerHash := txn.HashInner()
-
-	txn2 := *txn
-	txn2.Sigs = make([]cipher.Sig, len(txn.Sigs))
-	copy(txn2.Sigs, txn.Sigs)
-	txn2.In = make([]cipher.SHA256, len(txn.In))
-	copy(txn2.In, txn.In)
-	txn2.Out = make([]coin.TransactionOutput, len(txn.Out))
-	copy(txn2.Out, txn.Out)
-
-	if txnInnerHash != txn2.HashInner() {
-		logger.Panic("copyTransaction copy broke InnerHash")
-	}
-	if txnHash != txn2.Hash() {
-		logger.Panic("copyTransaction copy broke Hash")
-	}
-
-	return &txn2
-}
-
-// SignTransaction signs a transaction. Specific inputs may be signed by specifying signIndexes.
-// If signIndexes is empty, all inputs will be signed.
-// The transaction should already have a valid header. The transaction may be partially signed,
-// but a valid existing signature cannot be overwritten.
-// Clients should avoid signing the same transaction multiple times.
-func (w *Wallet) SignTransaction(txn *coin.Transaction, signIndexes []int, uxOuts []coin.UxOut) (*coin.Transaction, error) {
-	signedTxn := copyTransaction(txn)
-	txnInnerHash := signedTxn.HashInner()
-
-	if txnInnerHash != signedTxn.InnerHash {
-		return nil, NewError(errors.New("Transaction inner hash does not match computed inner hash"))
-	}
-
-	if len(signedTxn.Sigs) == 0 {
-		return nil, NewError(errors.New("Transaction signatures array is empty"))
-	}
-	if signedTxn.IsFullySigned() {
-		return nil, NewError(errors.New("Transaction is fully signed"))
-	}
-
-	if len(signedTxn.In) == 0 {
-		return nil, NewError(errors.New("No transaction inputs to sign"))
-	}
-	if len(uxOuts) != len(signedTxn.In) {
-		return nil, errors.New("len(uxOuts) != len(txn.In)")
-	}
-	if err := validateSignIndexes(signIndexes, uxOuts); err != nil {
-		return nil, NewError(err)
-	}
-
-	// Build a mapping of addresses to the inputs that need to be signed
-	addrs := make(map[cipher.Address][]int)
-	if len(signIndexes) > 0 {
-		for _, in := range signIndexes {
-			if !signedTxn.Sigs[in].Null() {
-				return nil, NewError(fmt.Errorf("Transaction is already signed at index %d", in))
-			}
-			addrs[uxOuts[in].Body.Address] = append(addrs[uxOuts[in].Body.Address], in)
-		}
-	} else {
-		for i, o := range uxOuts {
-			if !signedTxn.Sigs[i].Null() {
-				continue
-			}
-			addrs[o.Body.Address] = append(addrs[o.Body.Address], i)
-		}
-	}
-
-	// Check that the wallet has all addresses needed for signing
-	toSign := make(map[int][]int)
-	for i, e := range w.Entries {
-		if len(toSign) == len(addrs) {
-			break
-		}
-		addr := e.SkycoinAddress()
-		if x, ok := addrs[addr]; ok {
-			toSign[i] = x
-		}
-	}
-
-	if len(toSign) != len(addrs) {
-		return nil, NewError(errors.New("Wallet cannot sign all requested inputs"))
-	}
-
-	// Sign the selected inputs
-	for k, v := range toSign {
-		for _, x := range v {
-			if !signedTxn.Sigs[x].Null() {
-				return nil, NewError(fmt.Errorf("Transaction is already signed at index %d", x))
-			}
-			if err := signedTxn.SignInput(w.Entries[k].Secret, x); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	if err := signedTxn.UpdateHeader(); err != nil {
-		return nil, err
-	}
-
-	// Sanity check
-	if txnInnerHash != signedTxn.HashInner() {
-		err := errors.New("Transaction inner hash modified in the process of signing")
-		logger.Critical().WithError(err).Error()
-		return nil, err
-	}
-
-	return signedTxn, nil
-}
-
-// CreateTransaction creates and signs a transaction based upon CreateTransactionParams.
-// Set the password as nil if the wallet is not encrypted, otherwise the password must be provided.
-// NOTE: Caller must ensure that auxs correspond to params.Wallet.Addresses and params.Wallet.UxOuts options
-// Outputs to spend are chosen from the pool of outputs provided.
-// The outputs are chosen by the following procedure:
-//   - All outputs are merged into one list and are sorted coins highest, hours lowest, with the hash as a tiebreaker
-//   - Outputs are chosen from the beginning of this list, until the requested amount of coins is met.
-//     If hours are also specified, selection continues until the requested amount of hours are met.
-//   - If the total amount of coins in the chosen outputs is exactly equal to the requested amount of coins,
-//     such that there would be no change output but hours remain as change, another output will be chosen to create change,
-//     if the coinhour cost of adding that output is less than the coinhours that would be lost as change
-// If receiving hours are not explicitly specified, hours are allocated amongst the receiving outputs proportional to the number of coins being sent to them.
-// If the change address is not specified, the address whose bytes are lexically sorted first is chosen from the owners of the outputs being spent.
-func (w *Wallet) CreateTransaction(p CreateTransactionParams, auxs coin.AddressUxOuts, headTime uint64) (*coin.Transaction, []UxBalance, error) {
-	if err := p.Validate(); err != nil {
-		return nil, nil, err
-	}
-
-	if p.Wallet.ID != w.Filename() {
-		return nil, nil, NewError(errors.New("p.Wallet.ID does not match wallet"))
-	}
-
-	if w.IsEncrypted() {
-		return nil, nil, ErrWalletEncrypted
-	}
-
-	entriesMap := make(map[cipher.Address]Entry)
-	for a := range auxs {
-		// Check that auxs does not contain addresses that are not known to this wallet
-		e, ok := w.GetEntry(a)
-		if !ok {
-			return nil, nil, ErrUnknownAddress
-		}
-		entriesMap[e.SkycoinAddress()] = e
-	}
-
-	txn := &coin.Transaction{}
-
-	// Determine which unspents to spend
-	uxa := auxs.Flatten()
-
-	uxb, err := NewUxBalances(headTime, uxa)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Reverse lookup set to recover the inputs
-	uxbMap := make(map[cipher.SHA256]UxBalance, len(uxb))
-	for _, u := range uxb {
-		if _, ok := uxbMap[u.Hash]; ok {
-			return nil, nil, errors.New("Duplicate UxBalance in array")
-		}
-		uxbMap[u.Hash] = u
-	}
-
-	// calculate total coins and minimum hours to send
-	var totalOutCoins uint64
-	var requestedHours uint64
-	for _, to := range p.To {
-		totalOutCoins, err = mathutil.AddUint64(totalOutCoins, to.Coins)
-		if err != nil {
-			return nil, nil, NewError(fmt.Errorf("total output coins error: %v", err))
-		}
-
-		requestedHours, err = mathutil.AddUint64(requestedHours, to.Hours)
-		if err != nil {
-			return nil, nil, NewError(fmt.Errorf("total output hours error: %v", err))
-		}
-	}
-
-	// Use the MinimizeUxOuts strategy, to use least possible uxouts
-	// this will allow more frequent spending
-	// we don't need to check whether we have sufficient balance beforehand as ChooseSpends already checks that
-	spends, err := ChooseSpendsMinimizeUxOuts(uxb, totalOutCoins, requestedHours)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// calculate total coins and hours in spends
-	var totalInputCoins uint64
-	var totalInputHours uint64
-	var toSign []cipher.SecKey
-	if !p.Unsigned {
-		toSign = make([]cipher.SecKey, len(spends))
-	}
-	for i, spend := range spends {
-		totalInputCoins, err = mathutil.AddUint64(totalInputCoins, spend.Coins)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		totalInputHours, err = mathutil.AddUint64(totalInputHours, spend.Hours)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		entry, ok := entriesMap[spend.Address]
-		if !ok {
-			return nil, nil, fmt.Errorf("spend address %s not found in entriesMap", spend.Address.String())
-		}
-
-		if !p.Unsigned {
-			toSign[i] = entry.Secret
-		}
-		if err := txn.PushInput(spend.Hash); err != nil {
-			logger.Critical().WithError(err).Error("PushInput failed")
-			return nil, nil, err
-		}
-	}
-
-	feeHours := fee.RequiredFee(totalInputHours, params.UserVerifyTxn.BurnFactor)
-	if feeHours == 0 {
-		return nil, nil, fee.ErrTxnNoFee
-	}
-	remainingHours := totalInputHours - feeHours
-
-	switch p.HoursSelection.Type {
-	case HoursSelectionTypeManual:
-		txn.Out = append(txn.Out, p.To...)
-
-	case HoursSelectionTypeAuto:
-		var addrHours []uint64
-
-		switch p.HoursSelection.Mode {
-		case HoursSelectionModeShare:
-			// multiply remaining hours after fee burn with share factor
-			hours, err := mathutil.Uint64ToInt64(remainingHours)
-			if err != nil {
-				return nil, nil, err
-			}
-
-			allocatedHoursInt := p.HoursSelection.ShareFactor.Mul(decimal.New(hours, 0)).IntPart()
-			allocatedHours, err := mathutil.Int64ToUint64(allocatedHoursInt)
-			if err != nil {
-				return nil, nil, err
-			}
-
-			toCoins := make([]uint64, len(p.To))
-			for i, to := range p.To {
-				toCoins[i] = to.Coins
-			}
-
-			addrHours, err = DistributeCoinHoursProportional(toCoins, allocatedHours)
-			if err != nil {
-				return nil, nil, err
-			}
-		default:
-			return nil, nil, ErrInvalidHoursSelectionType
-		}
-
-		for i, out := range p.To {
-			out.Hours = addrHours[i]
-			txn.Out = append(txn.Out, out)
-		}
-
-	default:
-		return nil, nil, ErrInvalidHoursSelectionMode
-	}
-
-	totalOutHours, err := txn.OutputHours()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Make sure we have enough coins and coin hours
-	// If we don't, and we called ChooseSpends, then ChooseSpends has a bug, as it should have returned this error already
-	if totalOutCoins > totalInputCoins {
-		logger.WithError(ErrInsufficientBalance).Error("Insufficient coins after choosing spends, this should not occur")
-		return nil, nil, ErrInsufficientBalance
-	}
-
-	if totalOutHours > remainingHours {
-		logger.WithError(fee.ErrTxnInsufficientCoinHours).Error("Insufficient hours after choosing spends or distributing hours, this should not occur")
-		return nil, nil, fee.ErrTxnInsufficientCoinHours
-	}
-
-	// create change output
-	changeCoins := totalInputCoins - totalOutCoins
-	changeHours := remainingHours - totalOutHours
-
-	// If there are no change coins but there are change hours, try to add another
-	// input to save the change hours.
-	// This chooses an available input with the least number of coin hours;
-	// if the extra coin hour fee incurred by this additional input is less than
-	// the remaining coin hours, the input is added.
-	if changeCoins == 0 && changeHours > 0 {
-		// Find the output with the least coin hours
-		// If size of the fee for this output is less than the changeHours, add it
-		// Update changeCoins and changeHours
-		z := uxBalancesSub(uxb, spends)
-		sortSpendsHoursLowToHigh(z)
-		if len(z) > 0 {
-			extra := z[0]
-
-			// Calculate the new hours being spent
-			newTotalHours, err := mathutil.AddUint64(totalInputHours, extra.Hours)
-			if err != nil {
-				return nil, nil, err
-			}
-
-			// Calculate the new fee for this new amount of hours
-			newFee := fee.RequiredFee(newTotalHours, params.UserVerifyTxn.BurnFactor)
-			if newFee < feeHours {
-				err := errors.New("updated fee after adding extra input for change is unexpectedly less than it was initially")
-				logger.WithError(err).Error()
-				return nil, nil, err
-			}
-
-			// If the cost of adding this extra input is less than the amount of change hours we
-			// can save, use the input
-			additionalFee := newFee - feeHours
-			if additionalFee < changeHours {
-				changeCoins = extra.Coins
-
-				if extra.Hours < additionalFee {
-					err := errors.New("calculated additional fee is unexpectedly higher than the extra input's hours")
-					logger.WithError(err).Error()
-					return nil, nil, err
-				}
-
-				additionalHours := extra.Hours - additionalFee
-				changeHours, err = mathutil.AddUint64(changeHours, additionalHours)
-				if err != nil {
-					return nil, nil, err
-				}
-
-				entry, ok := entriesMap[extra.Address]
-				if !ok {
-					return nil, nil, fmt.Errorf("extra spend address %s not found in entriesMap", extra.Address.String())
-				}
-
-				if !p.Unsigned {
-					toSign = append(toSign, entry.Secret)
-				}
-				if err := txn.PushInput(extra.Hash); err != nil {
-					logger.Critical().WithError(err).Error("PushInput failed")
-					return nil, nil, err
-				}
-			}
-		}
-	}
-
-	// With auto share mode, if there are leftover hours and change couldn't be force-added,
-	// recalculate that share ratio at 100%
-	if changeCoins == 0 && changeHours > 0 && p.HoursSelection.Type == HoursSelectionTypeAuto && p.HoursSelection.Mode == HoursSelectionModeShare {
-		oneDecimal := decimal.New(1, 0)
-		if p.HoursSelection.ShareFactor.Equal(oneDecimal) {
-			return nil, nil, errors.New("share factor is 1.0 but changeHours > 0 unexpectedly")
-		}
-		p.HoursSelection.ShareFactor = &oneDecimal
-		return w.CreateTransaction(p, auxs, headTime)
-	}
-
-	if changeCoins > 0 {
-		var changeAddress cipher.Address
-		if p.ChangeAddress != nil {
-			changeAddress = *p.ChangeAddress
-		} else {
-			// Choose a change address from the unspent outputs
-			// Sort spends by address, comparing bytes, and use the first
-			// This provides deterministic change address selection from a set of unspent outputs
-			if len(spends) == 0 {
-				return nil, nil, errors.New("spends is unexpectedly empty when choosing an automatic change address")
-			}
-
-			addressBytes := make([][]byte, len(spends))
-			for i, s := range spends {
-				addressBytes[i] = s.Address.Bytes()
-			}
-
-			sort.Slice(addressBytes, func(i, j int) bool {
-				return bytes.Compare(addressBytes[i], addressBytes[j]) < 0
-			})
-
-			var err error
-			changeAddress, err = cipher.AddressFromBytes(addressBytes[0])
-			if err != nil {
-				logger.Critical().WithError(err).Error("cipher.AddressFromBytes failed for change address converted to bytes")
-				return nil, nil, err
-			}
-		}
-
-		if err := txn.PushOutput(changeAddress, changeCoins, changeHours); err != nil {
-			logger.Critical().WithError(err).Error("PushOutput failed")
-			return nil, nil, err
-		}
-	}
-
-	if p.Unsigned {
-		txn.Sigs = make([]cipher.Sig, len(txn.In))
-	} else {
-		txn.SignInputs(toSign)
-	}
-
-	// Attempt to zero secret keys from memory
-	var emptyKey cipher.SecKey
-	for i := range toSign {
-		copy(toSign[i][:], emptyKey[:])
-	}
-
-	if err := txn.UpdateHeader(); err != nil {
-		logger.Critical().WithError(err).Error("txn.UpdateHeader failed")
-		return nil, nil, err
-	}
-
-	inputs := make([]UxBalance, len(txn.In))
-	for i, h := range txn.In {
-		uxBalance, ok := uxbMap[h]
-		if !ok {
-			return nil, nil, errors.New("Created transaction's input is not in the UxBalanceSet, this should not occur")
-		}
-		inputs[i] = uxBalance
-	}
-
-	if err := verifyCreatedTransactionInvariants(p, txn, inputs); err != nil {
-		logger.Critical().WithError(err).Error("CreateTransaction created transaction that violates invariants, aborting")
-		return nil, nil, fmt.Errorf("Created transaction that violates invariants, this is a bug: %v", err)
-	}
-
-	return txn, inputs, nil
-}
-
-// verifyCreatedTransactionInvariants checks that the transaction that was created matches expectations.
-// Does not call visor verification methods because that causes import cycle.
-// daemon.Gateway checks that the transaction passes additional visor verification methods.
-func verifyCreatedTransactionInvariants(p CreateTransactionParams, txn *coin.Transaction, inputs []UxBalance) error {
-	for _, o := range txn.Out {
-		// No outputs should be sent to the null address
-		if o.Address.Null() {
-			return errors.New("Output address is null")
-		}
-
-		if o.Coins == 0 {
-			return errors.New("Output coins is 0")
-		}
-	}
-
-	if len(txn.Out) != len(p.To) && len(txn.Out) != len(p.To)+1 {
-		return errors.New("Transaction has unexpected number of outputs")
-	}
-
-	for i, o := range txn.Out[:len(p.To)] {
-		if o.Address != p.To[i].Address {
-			return errors.New("Output address does not match requested address")
-		}
-
-		if o.Coins != p.To[i].Coins {
-			return errors.New("Output coins does not match requested coins")
-		}
-
-		if p.To[i].Hours != 0 && o.Hours != p.To[i].Hours {
-			return errors.New("Output hours does not match requested hours")
-		}
-	}
-
-	if len(txn.Sigs) != len(txn.In) {
-		return errors.New("Number of signatures does not match number of inputs")
-	}
-
-	if p.Unsigned {
-		if !txn.IsFullyUnsigned() {
-			return errors.New("Transaction is not fully unsigned")
-		}
-	} else {
-		if !txn.IsFullySigned() {
-			return errors.New("Transaction is not fully signed")
-		}
-	}
-
-	if len(txn.In) != len(inputs) {
-		return errors.New("Number of UxOut inputs does not match number of transaction inputs")
-	}
-
-	for i, h := range txn.In {
-		if inputs[i].Hash != h {
-			return errors.New("Transaction input hash does not match UxOut inputs hash")
-		}
-	}
-
-	inputsMap := make(map[cipher.SHA256]struct{}, len(inputs))
-
-	for _, i := range inputs {
-		if i.Hours < i.InitialHours {
-			return errors.New("Calculated input hours are unexpectedly less than the initial hours")
-		}
-
-		if i.SrcTransaction.Null() {
-			return errors.New("Input's source transaction is a null hash")
-		}
-
-		if i.Hash.Null() {
-			return errors.New("Input's hash is a null hash")
-		}
-
-		if _, ok := inputsMap[i.Hash]; ok {
-			return errors.New("Duplicate input in array")
-		}
-
-		inputsMap[i.Hash] = struct{}{}
-	}
-
-	var inputHours uint64
-	for _, i := range inputs {
-		var err error
-		inputHours, err = mathutil.AddUint64(inputHours, i.Hours)
-		if err != nil {
-			return err
-		}
-	}
-
-	var outputHours uint64
-	for _, i := range txn.Out {
-		var err error
-		outputHours, err = mathutil.AddUint64(outputHours, i.Hours)
-		if err != nil {
-			return err
-		}
-	}
-
-	if inputHours < outputHours {
-		return errors.New("Total input hours is less than the output hours")
-	}
-
-	if inputHours-outputHours < fee.RequiredFee(inputHours, params.UserVerifyTxn.BurnFactor) {
-		return errors.New("Transaction will not satisy required fee")
-	}
-
-	return nil
-}
-
-// DistributeSpendHours calculates how many coin hours to transfer to the change address and how
-// many to transfer to each of the other destination addresses.
-// Input hours are split by BurnFactor (rounded down) to meet the fee requirement.
-// The remaining hours are split in half, one half goes to the change address
-// and the other half goes to the destination addresses.
-// If the remaining hours are an odd number, the change address gets the extra hour.
-// If the amount assigned to the destination addresses is not perfectly divisible by the
-// number of destination addresses, the extra hours are distributed to some of these addresses.
-// Returns the number of hours to send to the change address,
-// an array of length nAddrs with the hours to give to each destination address,
-// and a sum of these values.
-func DistributeSpendHours(inputHours, nAddrs uint64, haveChange bool) (uint64, []uint64, uint64) {
-	feeHours := fee.RequiredFee(inputHours, params.UserVerifyTxn.BurnFactor)
-	remainingHours := inputHours - feeHours
-
-	var changeHours uint64
-	if haveChange {
-		// Split the remaining hours between the change output and the other outputs
-		changeHours = remainingHours / 2
-
-		// If remainingHours is an odd number, give the extra hour to the change output
-		if remainingHours%2 == 1 {
-			changeHours++
-		}
-	}
-
-	// Distribute the remaining hours equally amongst the destination outputs
-	remainingAddrHours := remainingHours - changeHours
-	addrHoursShare := remainingAddrHours / nAddrs
-
-	// Due to integer division, extra coin hours might remain after dividing by len(toAddrs)
-	// Allocate these extra hours to the toAddrs
-	addrHours := make([]uint64, nAddrs)
-	for i := range addrHours {
-		addrHours[i] = addrHoursShare
-	}
-
-	extraHours := remainingAddrHours - (addrHoursShare * nAddrs)
-	i := 0
-	for extraHours > 0 {
-		addrHours[i] = addrHours[i] + 1
-		i++
-		extraHours--
-	}
-
-	// Assert that the hour calculation is correct
-	var spendHours uint64
-	for _, h := range addrHours {
-		spendHours += h
-	}
-	spendHours += changeHours
-	if spendHours != remainingHours {
-		logger.Panicf("spendHours != remainingHours (%d != %d), calculation error", spendHours, remainingHours)
-	}
-
-	return changeHours, addrHours, spendHours
-}
-
-// DistributeCoinHoursProportional distributes hours amongst coins proportional to the coins amount
-func DistributeCoinHoursProportional(coins []uint64, hours uint64) ([]uint64, error) {
-	if len(coins) == 0 {
-		return nil, errors.New("DistributeCoinHoursProportional coins array must not be empty")
-	}
-
-	coinsInt := make([]*big.Int, len(coins))
-
-	var total uint64
-	for i, c := range coins {
-		if c == 0 {
-			return nil, errors.New("DistributeCoinHoursProportional coins array has a zero value")
-		}
-
-		var err error
-		total, err = mathutil.AddUint64(total, c)
-		if err != nil {
-			return nil, err
-		}
-
-		cInt64, err := mathutil.Uint64ToInt64(c)
-		if err != nil {
-			return nil, err
-		}
-
-		coinsInt[i] = big.NewInt(cInt64)
-	}
-
-	totalInt64, err := mathutil.Uint64ToInt64(total)
-	if err != nil {
-		return nil, err
-	}
-	totalInt := big.NewInt(totalInt64)
-
-	hoursInt64, err := mathutil.Uint64ToInt64(hours)
-	if err != nil {
-		return nil, err
-	}
-	hoursInt := big.NewInt(hoursInt64)
-
-	var assignedHours uint64
-	addrHours := make([]uint64, len(coins))
-	for i, c := range coinsInt {
-		// Scale the ratio of coins to total coins proportionally by calculating
-		// (coins * totalHours) / totalCoins
-		// The remainder is truncated, remaining hours are appended after this
-		num := &big.Int{}
-		num.Mul(c, hoursInt)
-
-		fracInt := big.Int{}
-		fracInt.Div(num, totalInt)
-
-		if !fracInt.IsUint64() {
-			return nil, errors.New("DistributeCoinHoursProportional calculated fractional hours is not representable as a uint64")
-		}
-
-		fracHours := fracInt.Uint64()
-
-		addrHours[i] = fracHours
-		assignedHours, err = mathutil.AddUint64(assignedHours, fracHours)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if hours < assignedHours {
-		return nil, errors.New("DistributeCoinHoursProportional assigned hours exceeding input hours, this is a bug")
-	}
-
-	remainingHours := hours - assignedHours
-
-	if remainingHours > uint64(len(coins)) {
-		return nil, errors.New("DistributeCoinHoursProportional remaining hours exceed len(coins), this is a bug")
-	}
-
-	// For remaining hours lost due to fractional cutoff when scaling,
-	// first provide at least 1 coin hour to coins that were assigned 0.
-	i := 0
-	for remainingHours > 0 && i < len(coins) {
-		if addrHours[i] == 0 {
-			addrHours[i] = 1
-			remainingHours--
-		}
-		i++
-	}
-
-	// Then, assign the extra coin hours
-	i = 0
-	for remainingHours > 0 {
-		addrHours[i] = addrHours[i] + 1
-		remainingHours--
-		i++
-	}
-
-	return addrHours, nil
-}
-
-// UxBalance is an intermediate representation of a UxOut for sorting and spend choosing
-type UxBalance struct {
-	Hash           cipher.SHA256
-	BkSeq          uint64
-	Time           uint64
-	Address        cipher.Address
-	Coins          uint64
-	InitialHours   uint64
-	Hours          uint64
-	SrcTransaction cipher.SHA256
-}
-
-// NewUxBalances converts coin.UxArray to []UxBalance. headTime is required to calculate coin hours.
-func NewUxBalances(headTime uint64, uxa coin.UxArray) ([]UxBalance, error) {
-	uxb := make([]UxBalance, len(uxa))
-	for i, ux := range uxa {
-		b, err := NewUxBalance(headTime, ux)
-		if err != nil {
-			return nil, err
-		}
-		uxb[i] = b
-	}
-
-	return uxb, nil
-}
-
-// NewUxBalance converts coin.UxOut to UxBalance. headTime is required to calculate coin hours.
-func NewUxBalance(headTime uint64, ux coin.UxOut) (UxBalance, error) {
-	hours, err := ux.CoinHours(headTime)
-	if err != nil {
-		return UxBalance{}, err
-	}
-
-	return UxBalance{
-		Hash:           ux.Hash(),
-		BkSeq:          ux.Head.BkSeq,
-		Time:           ux.Head.Time,
-		Address:        ux.Body.Address,
-		Coins:          ux.Body.Coins,
-		InitialHours:   ux.Body.Hours,
-		Hours:          hours,
-		SrcTransaction: ux.Body.SrcTransaction,
-	}, nil
-}
-
-func uxBalancesSub(a, b []UxBalance) []UxBalance {
-	var x []UxBalance
-
-	bMap := make(map[cipher.SHA256]struct{}, len(b))
-	for _, i := range b {
-		bMap[i.Hash] = struct{}{}
-	}
-
-	for _, i := range a {
-		if _, ok := bMap[i.Hash]; !ok {
-			x = append(x, i)
-		}
-	}
-
-	return x
-}
-
-// ChooseSpendsMinimizeUxOuts chooses uxout spends to satisfy an amount, using the least number of uxouts
-//     -- PRO: Allows more frequent spending, less waiting for confirmations, useful for exchanges.
-//     -- PRO: When transaction is volume is higher, transactions are prioritized by fee/size. Minimizing uxouts minimizes size.
-//     -- CON: Would make the unconfirmed pool grow larger.
-// Users with high transaction frequency will want to use this so that they will not need to wait as frequently
-// for unconfirmed spends to complete before sending more.
-// Alternatively, or in addition to this, they should batch sends into single transactions.
-func ChooseSpendsMinimizeUxOuts(uxa []UxBalance, coins, hours uint64) ([]UxBalance, error) {
-	return ChooseSpends(uxa, coins, hours, sortSpendsCoinsHighToLow)
-}
-
-// sortSpendsCoinsHighToLow sorts uxout spends with highest balance to lowest
-func sortSpendsCoinsHighToLow(uxa []UxBalance) {
-	sort.Slice(uxa, makeCmpUxOutByCoins(uxa, func(a, b uint64) bool {
-		return a > b
-	}))
-}
-
-// ChooseSpendsMaximizeUxOuts chooses uxout spends to satisfy an amount, using the most number of uxouts
-// See the pros and cons of ChooseSpendsMinimizeUxOuts.
-// This should be the default mode, because this keeps the unconfirmed pool smaller which will allow
-// the network to scale better.
-func ChooseSpendsMaximizeUxOuts(uxa []UxBalance, coins, hours uint64) ([]UxBalance, error) {
-	return ChooseSpends(uxa, coins, hours, sortSpendsCoinsLowToHigh)
-}
-
-// sortSpendsCoinsLowToHigh sorts uxout spends with lowest balance to highest
-func sortSpendsCoinsLowToHigh(uxa []UxBalance) {
-	sort.Slice(uxa, makeCmpUxOutByCoins(uxa, func(a, b uint64) bool {
-		return a < b
-	}))
-}
-
-// sortSpendsHoursLowToHigh sorts uxout spends with lowest hours to highest
-func sortSpendsHoursLowToHigh(uxa []UxBalance) {
-	sort.Slice(uxa, makeCmpUxOutByHours(uxa, func(a, b uint64) bool {
-		return a < b
-	}))
-}
-
-func makeCmpUxOutByCoins(uxa []UxBalance, coinsCmp func(a, b uint64) bool) func(i, j int) bool {
-	// Sort by:
-	// coins highest or lowest depending on coinsCmp
-	//  hours lowest
-	//   oldest first
-	//    tie break with hash comparison
-	return func(i, j int) bool {
-		a := uxa[i]
-		b := uxa[j]
-
-		if a.Coins == b.Coins {
-			if a.Hours == b.Hours {
-				if a.BkSeq == b.BkSeq {
-					return cmpUxBalanceByUxID(a, b)
-				}
-				return a.BkSeq < b.BkSeq
-			}
-			return a.Hours < b.Hours
-		}
-		return coinsCmp(a.Coins, b.Coins)
-	}
-}
-
-func makeCmpUxOutByHours(uxa []UxBalance, hoursCmp func(a, b uint64) bool) func(i, j int) bool {
-	// Sort by:
-	// hours highest or lowest depending on hoursCmp
-	//  coins lowest
-	//   oldest first
-	//    tie break with hash comparison
-	return func(i, j int) bool {
-		a := uxa[i]
-		b := uxa[j]
-
-		if a.Hours == b.Hours {
-			if a.Coins == b.Coins {
-				if a.BkSeq == b.BkSeq {
-					return cmpUxBalanceByUxID(a, b)
-				}
-				return a.BkSeq < b.BkSeq
-			}
-			return a.Coins < b.Coins
-		}
-		return hoursCmp(a.Hours, b.Hours)
-	}
-}
-
-func cmpUxBalanceByUxID(a, b UxBalance) bool {
-	cmp := bytes.Compare(a.Hash[:], b.Hash[:])
-	if cmp == 0 {
-		logger.Panic("Duplicate UxOut when sorting")
-	}
-	return cmp < 0
-}
-
-// ChooseSpends chooses uxouts from a list of uxouts.
-// It first chooses the uxout with the most number of coins that has nonzero coinhours.
-// It then chooses uxouts with zero coinhours, ordered by sortStrategy
-// It then chooses remaining uxouts with nonzero coinhours, ordered by sortStrategy
-func ChooseSpends(uxa []UxBalance, coins, hours uint64, sortStrategy func([]UxBalance)) ([]UxBalance, error) {
-	if coins == 0 {
-		return nil, ErrZeroSpend
-	}
-
-	if len(uxa) == 0 {
-		return nil, ErrNoUnspents
-	}
-
-	for _, ux := range uxa {
-		if ux.Coins == 0 {
-			logger.Panic("UxOut coins are 0, can't spend")
-			return nil, errors.New("UxOut coins are 0, can't spend")
-		}
-	}
-
-	// Split UxBalances into those with and without hours
-	var nonzero, zero []UxBalance
-	for _, ux := range uxa {
-		if ux.Hours == 0 {
-			zero = append(zero, ux)
-		} else {
-			nonzero = append(nonzero, ux)
-		}
-	}
-
-	// Abort if there are no uxouts with non-zero coinhours, they can't be spent yet
-	if len(nonzero) == 0 {
-		return nil, fee.ErrTxnNoFee
-	}
-
-	// Sort uxouts with hours lowest to highest and coins highest to lowest
-	sortSpendsCoinsHighToLow(nonzero)
-
-	var have Balance
-	var spending []UxBalance
-
-	// Use the first nonzero output. This output will have the least hours possible
-	firstNonzero := nonzero[0]
-	if firstNonzero.Hours == 0 {
-		logger.Panic("balance has zero hours unexpectedly")
-		return nil, errors.New("balance has zero hours unexpectedly")
-	}
-
-	nonzero = nonzero[1:]
-
-	spending = append(spending, firstNonzero)
-
-	have.Coins += firstNonzero.Coins
-	have.Hours += firstNonzero.Hours
-
-	if have.Coins >= coins && fee.RemainingHours(have.Hours, params.UserVerifyTxn.BurnFactor) >= hours {
-		return spending, nil
-	}
-
-	// Sort uxouts without hours according to the sorting strategy
-	sortStrategy(zero)
-
-	for _, ux := range zero {
-		spending = append(spending, ux)
-
-		have.Coins += ux.Coins
-		have.Hours += ux.Hours
-
-		if have.Coins >= coins {
-			break
-		}
-	}
-
-	if have.Coins >= coins && fee.RemainingHours(have.Hours, params.UserVerifyTxn.BurnFactor) >= hours {
-		return spending, nil
-	}
-
-	// Sort remaining uxouts with hours according to the sorting strategy
-	sortStrategy(nonzero)
-
-	for _, ux := range nonzero {
-		spending = append(spending, ux)
-
-		have.Coins += ux.Coins
-		have.Hours += ux.Hours
-
-		if have.Coins >= coins && fee.RemainingHours(have.Hours, params.UserVerifyTxn.BurnFactor) >= hours {
-			return spending, nil
-		}
-	}
-
-	if have.Coins < coins {
-		return nil, ErrInsufficientBalance
-	}
-
-	return nil, ErrInsufficientHours
 }
